# ENH: Add ufunc for np.char.isalpha