--- a/numpy/core/code_generators/generate_umath.py
+++ b/numpy/core/code_generators/generate_umath.py
@@ -1268,8 +1268,9 @@ def make_arrays(funcdict):
             funcnames = ', '.join(funclist)
             signames = ', '.join(siglist)
             datanames = ', '.join(datalist)
-            code1list.append("static PyUFuncGenericFunction %s_functions[] = {%s};"
-                            % (name, funcnames))
+            code1list.append(
+                "static PyUFuncGenericFunction %s_functions[] = {%s};"
+                % (name, funcnames))
             code1list.append("static void * %s_data[] = {%s};"
                             % (name, datanames))
             code1list.append("static char %s_signatures[] = {%s};"