--- a/numpy/core/code_generators/ufunc_docstrings.py
+++ b/numpy/core/code_generators/ufunc_docstrings.py
@@ -4252,3 +4252,28 @@ add_newdoc('numpy.core.umath', 'bitwise_count',
           dtype=uint8)
 
     """)
+
+add_newdoc('numpy.core.umath', 'isalpha',
+    """
+    Returns true for each element if all characters in the string are
+    alphabetic and there is at least one character, false otherwise.
+
+    Calls :meth:`str.isalpha` element-wise.
+
+    For 8-bit strings, this method is locale-dependent.
+
+    Parameters
+    ----------
+    x : array_like of str or unicode
+    $PARAMS
+
+    Returns
+    -------
+    y : ndarray
+        Output array of bools
+        $OUT_SCALAR_1
+
+    See Also
+    --------
+    str.isalpha
+    """)