… benchmarks/benchmarks/bench_core.py