@lysnikolaou 2023-09-29
One problem with the current state of this PR that I can't really find a solution to is this:

The `isalpha` ufunc itself works correctly, both when called directly, and when called from the `np.char` module with an ndarray.

```python3
$ export PYTHONPATH="/Users/lysnikolaou/repos/python/numpy/build-install/usr/lib/python3.11/site-packages"
🐍 Launching Python with PYTHONPATH="/Users/lysnikolaou/repos/python/numpy/build-install/usr/lib/python3.11/site-packages"
$ /usr/bin/env python -P
Python 3.11.3 (tags/v3.11.3:f3909b8bc8, May 29 2023, 12:51:44) [Clang 14.0.3 (clang-1403.0.22.14.1)] on darwin
Type "help", "copyright", "credits" or "license" for more information.
>>> import numpy as np
>>> a = np.array([['abc', 'def'], ['world', ' he']])
>>> np.core.multi
np.core.multiarray np.core.multiply( 
>>> np.core.multiarray.isalpha(a)
array([[ True,  True],
       [ True, False]])
>>> np.char.isalpha(a)
array([[ True,  True],
       [ True, False]])
```

However, I'm getting a value error when passing a `chararray` to it.

```python3
$ export PYTHONPATH="/Users/lysnikolaou/repos/python/numpy/build-install/usr/lib/python3.11/site-packages"
🐍 Launching Python with PYTHONPATH="/Users/lysnikolaou/repos/python/numpy/build-install/usr/lib/python3.11/site-packages"
$ /usr/bin/env python -P
Python 3.11.3 (tags/v3.11.3:f3909b8bc8, May 29 2023, 12:51:44) [Clang 14.0.3 (clang-1403.0.22.14.1)] on darwin
Type "help", "copyright", "credits" or "license" for more information.
>>> import numpy as np
>>> a = np.char.array([['abc', 'def'], ['world', ' he']])
>>> np.core.multiarray.isalpha(a)
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/Users/lysnikolaou/repos/python/numpy/build-install/usr/lib/python3.11/site-packages/numpy/core/defchararray.py", line 2099, in __array_finalize__
    raise ValueError("Can only create a chararray from string data.")
ValueError: Can only create a chararray from string data.
```

It seems that the `bool` array returned from the ufuncs is somehow getting cast to a `chararray`, which is the input type. Is that something that should be happening? Am I missing something more obvious?