@seberg 2023-09-29
> Is that something that should be happening?

Yes, that should be happening.  Unlike a custom function, UFuncs will correctly preserve subclass information.

Could probably add something like (untested):
```
def __array_prepare__(self, arr, context=None):
    if arr.dtype in "US":
        return arr.view(type(self))
    return arr
```
to make it work, I doubt that has any negative side-effects.  Maybe `__array_wrap__` but I suspect `__array_prepare__` might be more targeeted even.  (Alternatively, there is a point why we don't just return a normal array when this happens.)