@lysnikolaou 2023-09-29
Is there a way to tell the ufunc to force a specific return type? Because `__array_finalize__` allows all of `SUbc`, I fear that it might break something if we change it in the `chararray` itself.