@lysnikolaou 2023-10-02
@mattip @seberg This appears to be okay now. I added an `__array_wrap__` which just returns the `arr` that's being passed in. I don't have a deep enough understanding of the ufunc mechanism yet to figure out if this is okay like this, so I'd like a review please.