@mattip 2023-10-02
Nice. I think the `__array_wrap__` is correct. Could you rerun the benchmarks?