@lysnikolaou 2023-10-02
> Could you rerun the benchmarks?

Similar results:

```
| Change   | Before [eabefa4e] <main>   | After [54ef4672] <string-ufuncs-isalpha>   |   Ratio | Benchmark (Parameter)                                   |
|----------|----------------------------|--------------------------------------------|---------|---------------------------------------------------------|
| -        | 7.64±0.04μs                | 2.30±0.1μs                                 |     0.3 | bench_core.NumPyChar.time_isalpha_small_list_big_string |
| -        | 1.34±0.02ms                | 5.15±0.2μs                                 |     0   | bench_core.NumPyChar.time_isalpha_big_list_small_string |
```