@lysnikolaou 2023-10-03
@ngoldbaum Is `stringdtype` currently broken on main? I'm trying to test this and building `stringdtype` fails both with this branch and with the numpy nightly build.