@ngoldbaum 2023-10-03
> I'm trying to test this and building stringdtype fails both with this branch and with the numpy nightly build

I don't see this when I build stringdtype using numpy's main branch or this PR. Can you open an issue over on the numpy-user-dtypes repo with more detail about what's happening on your setup? Maybe it's just a documentation thing but if it isn't I'd like to know what the problem is.

FWIW, this does break `isalpha` for `stringdtype` because the dtype doesn't have an implementation of the ufunc yet, but I think that's expected. No need for numpy to handle that I think, I can fix it in stringdtype once this is merged.