@lysnikolaou 2023-10-03
Sounds good. I'll try to recreate and will open a PR.

@mattip @seberg I've got a follow-up PR ready that adds support for `np.add`. Can we merge this so that I can open that PR?