@ngoldbaum 2023-10-04
I'm glad it was relatively straightforward to add support for string ufuncs besides the logic operators, looking forward to improved string operator performance! :rocket: