@lysnikolaou 2023-12-11
@charris This was the first one of the PRs adding ufuncs for `np.char` and we'd like to backport this to 1.x. Is it okay to tag you in all of the other PRs that need backporting as well? I'm not sure about the exact process here, so I'd need a few pointers.