@seberg 2023-12-11
While it seems nice that these are backportable.  I don't really think they should be backported for a 1.26.x bug-fix release.  These are quite large changes, and chances are some behavior is changed.